Distribute classes by modules
